backend/requirements.txt
fastapi
uvicorn[standard]
gigachat
pydantic-settings
sqlalchemy[asyncio]
aiosqlite
ruff
mcp
apscheduler>=3.10,<4
croniter>=1.3,<3
numpy
